Add k8s-external-secrets regional application
We introduce charts/region/k8s-external-secrets which deploys the
kubernetes-external-secret chart on regional clusters pointing to the
vault installed on the hub. Note that until we create a secret called
'vault-ca' in the k8s-external-secret namespace, the operator pod will
not be healthy as a mountpoint will be missing.

This depends on https://github.com/hybrid-cloud-patterns/common/pull/37
to work.

Tested in conjunction with the common PR and I correctly am able to get
the correct VAULT_ADDR configured:

  for i in $(oc -n k8s-external-secrets get pods -o jsonpath='{range .items[*].metadata}{.name}{end}'); do
    oc exec -n k8s-external-secrets -it "$i" -- env |grep VAULT_ADDR
  done
  VAULT_ADDR=vault-vault.apps.bandini-dc.blueprints.rhecoeng.com