Add a secret example in config-demo app
This set adds a secret inside the config-demo app.

The secret can be injected either manually in the vault container via:
  vault kv put secret/hub/config-demo secret=bar

Or one can create ~/values-secret.yaml:
secrets:
  config-demo:
    secret: test123

And then run ./common/scripts/ansible-push-vault-secrets.sh

Due to https://issues.redhat.com/browse/ACM-1208 the order of the the
commands is quite important. Namely the user has to run the following:
1. make install on hub
2. run make vault-init
3. run ./common/scripts/ansible-push-vault-secrets.sh (or push the
   config-demo/secret in the vault manually)
4. join remote cluster

With the above sequence it all works out of the box.
Should the sequence above not be respected (i.e. join the cluster before
unsealing the vault and adding the secret), the user would have to run
oc edit -n <remote-acm-cluster-name> policy/config-demo.configdemo-region-secret-policy
and modify the trigger-update annotation so that ACM will be forced to
refresh the policy after the hub lookup failed during deploy (because
the vault part was not ready).

See https://issues.redhat.com/browse/ACM-1208 for further info.