Require ACM 2.5
Let's switch to ACM 2.5 for multicloud gitops. This allows us to remove
the {{ hub + lookup }} from templates and switch to the more secure
{{ hub + fromSecret }}

Tested this and I correctly get the config-demo app + secret deployed on
the regional cluster.

Note: Once this is merge we *cannot* go back to ACM 2.4 because the
combo hub + fromSecret is not supported there and ACM will error out
with: "Error processing hub templates"