Add super linter
We disable the following linters to begin with:

  VALIDATE_BASH: false
  VALIDATE_JSCPD: false
  VALIDATE_KUBERNETES_KUBEVAL: false
  VALIDATE_YAML: false
  VALIDATE_ANSIBLE: false

Kubeval is already covered by us and we cannot use this one because it
does not have all of our OCP-specific schemas. JSCP needs to be disabled
because our is not yaml for the most part but helm templates which are
not recognized. Same with the YAML validator.
We can debate if we reenable shell later. In any case we should to it
only after we merge the vault utils ansible port.
Ansible is currently disabled and we can reenable it once we move to
ansible for vault stuff.

While we're at it we fix some markdown and some more obvious linting
issues.

Our main goal is to always have gitleaks running to be notified if a
key/password gets pushed accidentally. We can see that it is indeed
running:

2022-06-27 15:47:57 [INFO]   ---------------------------
2022-06-27 15:47:57 [INFO]   File:[/github/workspace/clustergroup/templates/namespaces.yaml]
2022-06-27 15:47:57 [INFO]    - File:[namespaces.yaml] was linted with [gitleaks] successfully
2022-06-27 15:47:57 [INFO]      - Command output:
------
    ○
    │╲
    │ ○
    ○ ░
    ░    gitleaks
3:47PMINF scan completed in 404.005µs
3:47PMINF no leaks found

Also add a 'super-linter' Makefile target to test linting locally
easily.