Add support for .disabled in imperative jobs
If a job was the "disabled: true" key in its definition
it won't be rendered in the CronJob definition.
The use case here is to be able to simply noop some jobs
without having to remove them fully.

Tested with:
  imperative:
    jobs:
    - name: foo
      playbook: ansible/playbooks/foo.yml

    - name: bar
      playbook: ansible/playbooks/bar.yml
      disabled: true

    - name: baz
      playbook: ansible/playbooks/baz.yml
      disabled: false

The above generates the following initContainers in the
CronJob definition (note how bar is skipped):

initContainers:
...
  - name: foo
    image: registry.redhat.io/ansible-automation-platform-22/ee-supported-rhel8:latest
    command:
    ...
      - ansible/playbooks/foo.yml
  - name: baz
    command:
    ...
      - ansible/playbooks/test.yml