Improve validate-origin
Currently we only run git-ls against the remote repo.
While this is helpful it has two drawbacks:
1) It does not check that the local branch exists remotely
This is a common error where an operator did the local changes in
a local branch but forgot to push it remotely. The only way
to catch this error is to logon to the main argo and see the error
about remote SHA not being able to be resolved.

2) It can be very verbose due to having many remote branches and
clutter the output a bit.

Let's tweak this by actually checking that the local branch exists
remotely and bailing out if it does not

Tested as follows:
- Local branch not pushed remotely

    make validate-origin
    Checking repo https://github.com/hybrid-cloud-patterns/common - branch ux-improvements
    ux-improvements not found in https://github.com/hybrid-cloud-patterns/common
    make: *** [Makefile:43: validate-origin] Error 1

- Local branch that also exists remotely:

    make validate-origin
    make -f common/Makefile validate-origin
    make[1]: Entering directory '/home/michele/Engineering/cloud-patterns/multicloud-gitops'
    Checking repo https://github.com/mbaldessari/multicloud-gitops.git - branch no-values-url
    https://github.com/mbaldessari/multicloud-gitops.git - no-values-url exist
    make[1]: Leaving directory '/home/michele/Engineering/cloud-patterns/multicloud-gitops'