Cleanup configManagementPlugins section in argocd.yaml
The current yaml is unreadable for any practical means.
Let's use the '|' yaml operator to join the lines
nicely. This is also done upstream [1]. This paves the
way to add some code that will skip non existing values
files.

Tested by me and Lester and we got a correctly
deployed manuela-tst (which is a user of the plugin systems)

[1] https://argo-cd.readthedocs.io/en/stable/user-guide/config-management-plugins/