WIP/Do not merge: vault as subchart
Due to https://github.com/helm/helm/issues/9136 we need to set the
toplevel vault.server.host helm parameter explicitely (as null, does not
work). Other approaches like patching the vault-helm chart to accept
'false' as route and actually set 'null' in that case have been explored
but they seemed to hacky.

With this approach we get our own chart and the values file is a lot
simpler.

This is just a patch to show the approach. If we agree on the approach,
I'll split the common/ bit out first and then only use the values-hub
approach.