Use UBI based images for ESO
Thsi changes switches to use UBI based images as reccomended by
Matthias.

Tested on MCG on hub and also checked the regional cluster is still
working correctly:

    oc get -n golang-external-secrets pods -o yaml |grep -i image:
	  image: ghcr.io/external-secrets/external-secrets:v0.6.0-ubi
	  image: ghcr.io/external-secrets/external-secrets:v0.6.0-ubi
	  image: ghcr.io/external-secrets/external-secrets:v0.6.0-ubi
	  image: ghcr.io/external-secrets/external-secrets:v0.6.0-ubi
	  image: ghcr.io/external-secrets/external-secrets:v0.6.0-ubi
	  image: ghcr.io/external-secrets/external-secrets:v0.6.0-ubi

Note that after this change, when we update the ESO subchart we also
need to tweak the version. That is why I added a README with the
reminder.