Move some of the oc calls into the legacy target
Namely HUBCLUSTER_APPS_DOMAIN and HUBCLUSTER_VERSION do not need to be
calculated at every command. So let's move them into the legacy target.
We do not need them in the operator-based deploy as those are calculated
by the operator anyways.

Tested both with 'make install' and 'make legacy-install' in MCG.