Drop legacy-install from common
We now only support the installation via the operator-based templates.
This way we avoid duplication and get a certain amount of flexibility
as the operator is in a better place to fetch information from the
cluster (clusterVersion, DNS, etc.)

As an added bonus we can drop most 'oc' calls that were only really
needed in the legacy install form.

When the old target is invoked, let's print an error message.