Add SkipDryRunOnMissingResource=true on ClusterPools, ClusterClaims and ManagedClusterSets
On a deploy using clusterPools the acm app will stay in error forever.
The reason for this is that the CRDs for the above objects won't exist
until the MultiClusterHub is installed. So Argo CD won't find the CRD
in the sync and will fail with the error the server could not find the
requested resource.

Add the "argocd.argoproj.io/sync-options: SkipDryRunOnMissingResource=true" on those resources, just like we do for Policies.

Co-Authored-By: Lester Claudio <claudiol@redhat.com>