Add ini_file support
With this support a user can specify ini_file instead of path or value
inside a secret's field. So the following is now possible:
 secrets:
  - name: aws
    fields:
    - name: aws_access_key_id
      ini_file: ~/.aws/credentials
      ini_key: aws_access_key_id
      # ini_section: default
    - name: aws_secret_access_key
      ini_file: ~/.aws/credentials
      ini_key: aws_secret_access_key
      # ini_section: default

This would read ~/.aws/credentials extract the key and place it in the
vault path (default is 'hub'). So the above would generate the
'secret/hub/aws' vault key with the the 'aws_access_key_id' and
'aws_secret_access_key' attributes containing the content of the keys
in the inifile.