Add trivial Hello World demo
This simple helm app is useful to demo a simple app that adds a route,
displays a helm variable and nothing else. I.e. this can work without
vault or secrets being involved.

This is useful when testing Hosted Argo stuff because the framework is
a bit more limited in capabilities in that context.

Tested it and it all worked as expected.

Note that we do not enable this by default, it's there in case but not
enabled by default like config-demo is.