Drop SSH_SOCK_MOUNT from pattern-util.sh
As reported by Jonny this might be problematic on some Mac OSX versions.
While the investigation still needs to be done, we can safely drop this
additional mount point because in the 'validate-origin' target we are
currently skipping the 'git ls-remote' call when running inside the
container.

We can reassess this at a later time and debug all these corner case,
but it will only make sense if we run the git ls-remote check in the
container as well.