Add more logic to `install` target
So this commit aims to tweak the `install` make target so that it can be
invoked whenever we need to tweak things like git branch or git repo via
CLI but the VP pattern has been installed via the OperatorHub.
Currently if you just run `make install` from the repo after the
operator has been installed via the Web UI, there will be an error like:
```
Error: rendered manifests contain a resource that already exists. Unable
to continue with install: Pattern "multicloud-gitops" in namespace
"openshift-operators" exists and cannot be imported into the current
release: invalid ownership metadata; label validation error: missing key
"app.kubernetes.io/managed-by": must be set to "Helm"; annotation
validation error: missing key "meta.helm.sh/release-name": must be set
to "multicloud-gitops"; annotation validation error: missing key
"meta.helm.sh/release-namespace": must be set to "default"
```

The problem is that the pattern object and the patterns operator
subscription are missing some annotations and labels and that confuses
helm install later on as it sees the object it needs to create but craps
itself as it has no knowledge of it.

A few approaches were investigated to solve this:
1. Tweak the operator to add annotations. This, besides being a bit
   fragile and convoluted, is not a viable path, because the operator
   would also have to tweak the subscription that was used to install
   itself, which seems grossly out of bounds.
2. Tweak `make install` to always use `helm template` and pipe it
   through `oc apply`. This quickly would become rather complex and
   fragile because of the way CRDs are managed in helm: we would have to
   first install the CRD, wait for its successful creation and only then
   do another `helm template ... | oc apply -f-`. This quickly becomes
   rather involved and fragile. Also it is really reimplementing all of
   the `helm install` logic.
3. Add a post-renderer script to strip the standard helm annotations
   (https://helm.sh/docs/chart_best_practices/labels/#standard-labels).
   Seemingly helm does not pass those annotations to the post renderer
   but adds them afterwards, because a --post-renderer pointing to a
   script that does `yq e 'del(.metadata.annotations)'` did not work

Given the above limits, I decided to tweak `make install` so that if the
CRD for the pattern is missing, it would just run the usual helm
install. If the CRD is present then either it has been installed via the
operator's UI or via `make install`. Either way, we just just run helm
template and pipe it to oc apply, so that we won't force the helm
standard annotations.

We have two invocations of helm template (one piped to oc apply
set-last-applied) so we avoid the following warning during the first
reapply call:

    Warning: resource patterns/multicloud-gitops is missing the
    kubectl.kubernetes.io/last-applied-configuration annotation which is
    required by oc apply. oc apply should only b e used on resources created
    declaratively by either oc create --save-config or oc apply. The missing
    annotation will be patched automatically.
    pattern.gitops.hybrid-cloud-patterns.io/multicloud-gitops configured
    Warning: resource subscriptions/patterns-operator is missing the
    kubectl.kubernetes.io/last-applied-configuration annotation which is
    required by oc apply. oc apply should o nly be used on resources created
    declaratively by either oc create --save-config or oc apply. The missing
    annotation will be patched automatically.
    subscription.operators.coreos.com/patterns-operator configured

Tested by installing a pattern via UI of the VP Operator, then locally ran
`make install` and observed the new git branch being applied in the
operator:

    Reapplying helm chart:
    pattern.gitops.hybrid-cloud-patterns.io/multicloud-gitops configured
    subscription.operators.coreos.com/patterns-operator configured
    pattern.gitops.hybrid-cloud-patterns.io/multicloud-gitops configured
    subscription.operators.coreos.com/patterns-operator configured
    make[1]: Leaving directory '/home/michele/Engineering/cloud-patterns/multicloud-gitops'

Tested also with a clean `make install` from scratch:

    Running helm:
    Release "multicloud-gitops" does not exist. Installing it now.
    NAME: multicloud-gitops
    LAST DEPLOYED: Wed Mar 15 12:14:49 2023
    NAMESPACE: default
    STATUS: deployed
    REVISION: 1
    TEST SUITE: None