Fix initcontainer race on spokes
initContainers on the clusterwide argo instances on spokes rely on a
configmap called "trusted-ca-bundle" that gets created by ACM in the
openshift-gitops namespace.

Since we have no explicit guarantees about ordering, it might happen
that the argocd object gets created by acm before the configmap. This
is problematic because the init container will start but just ignore
the missing configmap.

Solve this by splitting the ocp-gitops-policy in two and making sure
that both the openshift-gitops subscription and the trusted-ca-bundle
configmap are green before applying the next dependency.