Properly quote vault kv command
When using a password with a dollar, the oc exec commands run inside the
vault are not properly escaped, casing the dollar signed to be
interpreted by the shell inside the vault pod.

So a password like 'Y$yxn54&qXAxpUd2*yGH' will become 'Y&qXAxpUd2*yGH'
in the vault.

This is because the command that is being run ends up being:

    oc exec -n vault vault-0 -i -- sh -c "vault kv patch -mount=secret global/mysecret dollar=Y$yxn54&qXAxpUd2*yGH"

The `$yxn54` will be interpreted by the shell inside vault.

Let's fix this by running a properly escaped command:

    oc exec -n vault vault-0 -i -- sh -c "vault kv patch -mount=secret global/mysecret dollar='"'Y$yxn54&qXAxpUd2*yGH'"'"

Reported-By: Chris Butler <chris.butler@redhat.com>